move helper function to proper class
